requirements.txt
asgiref==3.8.1
boto3==1.37.15
botocore==1.37.15
Django==5.1.7
django-colorfield==0.12.0
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-storages==1.14.5
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.1.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0